requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
et-xmlfile==1.1.0
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
pylightxl==1.61
python-dotenv==1.0.1
pywin32==306
yarl==1.9.4
